refactor: 컴포넌트 폴더 정리 및 재분류
